Move parts from MainLoop to AbstractEventLoop
This is preparation for implementation of GLib event loop to
Simpleline.
