+++ simpleline/event_loop/__init__.py
@@ -25,6 +25,12 @@ from abc import ABCMeta, abstractmethod
 from collections import namedtuple
 
 from simpleline.errors import SimplelineError
+from simpleline.event_loop.ticket_machine import TicketMachine
+from simpleline.logging import get_simpleline_logger
+
+log = get_simpleline_logger()
+
+__all__ = ["AbstractEventLoop", "AbstractSignal", "ExitMainLoop"]
 
 QuitCallback = namedtuple("QuitCallback", ["callback", "args"])
 
@@ -37,10 +43,13 @@ class ExitMainLoop(SimplelineError):
 class AbstractEventLoop(metaclass=ABCMeta):
 
     def __init__(self):
-        self._quit_callback = None
         super().__init__()
+        self._handlers = {}
+        self._processed_signals = TicketMachine()
+        self._quit_callback = None
+        # end most inner loop politely by setting to False
+        self._run_loop = True
 
-    @abstractmethod
     def register_signal_handler(self, signal, callback, data=None):
         """Register a callback which will be called when message "event"
         is encountered during process_events.
@@ -49,16 +58,18 @@ class AbstractEventLoop(metaclass=ABCMeta):
         - the received message in the form of (type, [arguments])
         - the data registered with the handler
 
-        :param signal: signal class we want to react on
-        :type signal: class based on the simpleline.event_loop.AbstractSignal class
+        :param signal: Signal class we want to react on.
+        :type signal: Class based on the simpleline.event_loop.AbstractSignal class.
 
-        :param callback: the callback function
+        :param callback: The callback function.
         :type callback: func(event_message, data)
 
-        :param data: optional data to pass to callback
-        :type data: anything
+        :param data: Optional data to pass to callback.
+        :type data: Anything.
         """
-        pass
+        if signal not in self._handlers:
+            self._handlers[signal] = []
+        self._handlers[signal].append(EventHandler(callback, data))
 
     @abstractmethod
     def register_signal_source(self, signal_source):
@@ -73,15 +84,15 @@ class AbstractEventLoop(metaclass=ABCMeta):
     def enqueue_signal(self, signal):
         """Enqueue new event for processing.
 
-        :param signal: signal which you want to add to the event queue for processing
-        :type signal: instance based on AbstractEvent class
+        :param signal: Signal which you want to add to the event queue for processing.
+        :type signal: Instance based on AbstractEvent class.
         """
-        pass
+        log.debug("New signal %s enqueued with source %s", signal, signal.source.__class__.__name__)
 
     @abstractmethod
     def run(self):
         """Starts the event loop."""
-        pass
+        log.debug("Starting main loop")
 
     @abstractmethod
     def execute_new_loop(self, signal):
@@ -89,10 +100,10 @@ class AbstractEventLoop(metaclass=ABCMeta):
 
         This is required for processing a modal screens.
 
-        :param signal: signal passed to the new event loop
-        :type signal: `AbstractSignal` based class
+        :param signal: Signal passed to the new event loop.
+        :type signal: The `AbstractSignal` based class.
         """
-        pass
+        log.debug("Executing inner loop")
 
     @abstractmethod
     def close_loop(self):
@@ -100,7 +111,7 @@ class AbstractEventLoop(metaclass=ABCMeta):
 
         Close an event loop created by the `execute_new_loop()` method.
         """
-        pass
+        log.debug("Closing inner loop")
 
     @abstractmethod
     def process_signals(self, return_after=None):
@@ -130,6 +141,43 @@ class AbstractEventLoop(metaclass=ABCMeta):
         :type args: Anything.
         """
         self._quit_callback = QuitCallback(callback, args)
+
+    def _register_wait_on_signal(self, wait_on_signal):
+        """Register process waiting on signal `wait_on_signal` and return id for later checking.
+
+        ID is returned which is then used in the `self._check_if_signal_processed()` method.
+
+        :param wait_on_signal: Signal we are waiting for.
+        :type wait_on_signal: Class based on `simpleline.event_loop.AbstractSignal`.
+        """
+        return self._processed_signals.take_ticket(wait_on_signal.__name__)
+
+    def _mark_signal_processed(self, signal):
+        """Mark that processes waiting on this signal that they are able to go.
+
+        :param signal: Signal which was processed.
+        :type signal: Class based on `simpleline.event_loop.AbstractSignal`.
+        """
+        self._processed_signals.mark_line_to_go(signal.__class__.__name__)
+
+    def _check_if_signal_processed(self, wait_on_signal, unique_id):
+        """Check if the signal was processed.
+
+        :param wait_on_signal: Signal the process is waiting for.
+        :type wait_on_signal: Class based on `simpleline.event_loop.AbstractSignal`.
+
+        :param unique_id: Unique id returned by the `self._register_wait_on_signal()` method.
+        :type unique_id: int
+        """
+        return self._processed_signals.check_ticket(wait_on_signal.__name__, unique_id)
+
+
+class EventHandler(object):
+    """Data class to save event handlers."""
+
+    def __init__(self, callback, data):
+        self.callback = callback
+        self.data = data
 
 
 class AbstractSignal(metaclass=ABCMeta):
